refactor: make groups logic one page
